Move some functions from clang CodeGen to clang AST
These functions are defined in HlslTypes.h in clang AST so they
should have implementations in libclangAST.a.
